ho: Don't handle Namespace updates on Node
This commit removes the Namespace event handler from the Node
controller. Nodes don't need to respond to these events as the
master will update pods with the necessary annotations if a
namespace changes.
